Created tests to define scope
