+++ src/main/java/org/openrewrite/java/testing/assertj/AdoptAssertJNumberAssert.java
@@ -1,0 +1,15 @@
+package org.openrewrite.java.testing.assertj;
+
+import org.openrewrite.Recipe;
+
+public class AdoptAssertJNumberAssert extends Recipe {
+    @Override
+    public String getDisplayName() {
+        return "Display Name";
+    }
+
+    @Override
+    public String getDescription() {
+        return "...";
+    }
+}
